lobby-check: fix two real bugs found while testing against a live photo_id
- photo_id was a 32-bit Integer column; real photo_ids (e.g.
  1787248984204) exceed Postgres's int32 range and the insert failed with
  "integer out of range" the first time a real id was submitted. Changed to
  BigInteger in both the model and the migration.
- fetch_image rejected application/octet-stream, which is S3's default
  Content-Type for an object uploaded without one explicitly set -- this
  bucket's real objects hit that case, so a genuinely valid .jpg was being
  rejected before ever reaching Bedrock. octet-stream is now accepted; the
  actual image-or-not decision is left to image_framing's PIL decode, which
  is what this check exists to backstop, not replace.

Verified end-to-end against the real local stack with a real production
image (mm-intelligence.s3.amazonaws.com/.../1787248984204.jpg): full
submit -> fetch -> Bedrock extraction -> completed round trip, correct
movie_title/material_type/material_quantity/material_condition + all four
confidences, and a matching LlmCallLog row (task_type=lobby_check,
call_path=bedrock_converse, cost_usd=0.00155, decision=AUTO_ACCEPT).

+++ backend/alembic/versions/d5e6f7a8b9c0_add_lobby_check_tables.py
@@ -44,7 +44,9 @@ def upgrade() -> None:
             'lobbycheckrow',
             sa.Column('id', sa.Integer(), nullable=False),
             sa.Column('job_id', sa.String(), nullable=False),
-            sa.Column('photo_id', sa.Integer(), nullable=False),
+            # BigInteger -- real photo_ids (e.g. 1787248984204) exceed
+            # Postgres's 32-bit Integer range.
+            sa.Column('photo_id', sa.BigInteger(), nullable=False),
             sa.Column('image_url', sa.String(), nullable=False),
             sa.Column('input_json', sa.String(), nullable=False),
             sa.Column('status', sa.String(), nullable=False, server_default='pending'),
